hierophant api and artifact_store refactor
