+++ hierophant/hierophant/src/artifact_store/mod.rs
@@ -1,0 +1,6 @@
+mod artifact_uri;
+mod client;
+mod command;
+mod store;
+
+pub use {artifact_uri::ArtifactUri, client::ArtifactStoreClient};
